full JS -> TS
