+++ package.json
@@ -6,7 +6,7 @@
     "mcsc": "bin/mcsc"
   },
   "scripts": {
-    "build": "npx tsc"
+    "build": "cp -r src/template dist/template && npx tsc"
   },
   "publishConfig": {
     "access": "public",
@@ -35,6 +35,8 @@
     "inquirer": "^8.1.1",
     "mcservercreator": "^1.0.1",
     "path": "^0.12.7",
+    "tsc": "^2.0.4",
+    "typescript": "^4.9.4",
     "yargs": "^17.5.1"
   },
   "devDependencies": {
